move: package foldering & naming
변경사항
- utils -> common
- presentation 내 utils 패키지 추가
- ui -> designsystem
- 일부 model 내 data class -> presentation - play - domain - model 이동
