+++ app/src/main/java/com/jbrunoo/digitink/common/PlayMode.kt
@@ -1,4 +1,4 @@
-package com.jbrunoo.digitink.utils
+package com.jbrunoo.digitink.common
 
 enum class PlayMode {
     NORMAL,
