+++ app/src/main/java/com/jbrunoo/digitink/designsystem/component/BigText.kt
@@ -1,4 +1,4 @@
-package com.jbrunoo.digitink.presentation.component
+package com.jbrunoo.digitink.designsystem.component
 
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
