+++ app/src/main/java/com/jbrunoo/digitink/designsystem/component/DiButton.kt
@@ -1,4 +1,4 @@
-package com.jbrunoo.digitink.presentation.component
+package com.jbrunoo.digitink.designsystem.component
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.RowScope
