+++ app/src/main/java/com/jbrunoo/digitink/presentation/play/component/PlayBoard.kt
@@ -39,13 +39,13 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.jbrunoo.digitink.domain.model.DrawPath
-import com.jbrunoo.digitink.domain.model.QnaWithPath
-import com.jbrunoo.digitink.presentation.play.model.PlayBoardState
-import com.jbrunoo.digitink.presentation.play.model.rememberPlayBoardState
-import com.jbrunoo.digitink.utils.drawCorrectIndicator
-import com.jbrunoo.digitink.utils.drawIncorrectIndicator
-import com.jbrunoo.digitink.utils.drawUserPaths
+import com.jbrunoo.digitink.presentation.play.domain.model.DrawPath
+import com.jbrunoo.digitink.presentation.play.domain.model.QnaWithPath
+import com.jbrunoo.digitink.presentation.play.domain.model.PlayBoardState
+import com.jbrunoo.digitink.presentation.play.domain.model.rememberPlayBoardState
+import com.jbrunoo.digitink.presentation.utils.extension.drawCorrectIndicator
+import com.jbrunoo.digitink.presentation.utils.extension.drawIncorrectIndicator
+import com.jbrunoo.digitink.presentation.utils.extension.drawUserPaths
 import kotlinx.coroutines.async
 import kotlinx.coroutines.launch
 import timber.log.Timber
