+++ app/src/main/java/com/jbrunoo/digitink/presentation/play/infinite/InfinitePlayScreen.kt
@@ -11,7 +11,7 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.jbrunoo.digitink.presentation.play.component.LifeLayout
 import com.jbrunoo.digitink.presentation.play.component.PlayBoard
-import com.jbrunoo.digitink.presentation.play.model.rememberPlayBoardState
+import com.jbrunoo.digitink.presentation.play.domain.model.rememberPlayBoardState
 
 @Composable
 fun InfinitePlayScreen(
